Refactor PathItemTypeAttribute for shared metadata and correct webhook classification
