+++ src/Grpc/JsonTranscoding/test/testassets/Sandbox/Services/Metadata/PathItemTypeAttribute.cs
@@ -1,0 +1,15 @@
+namespace Sandbox.Services.Metadata
+{
+    public enum PathItemType
+    {
+        Standard,
+        Webhook
+    }
+
+    [AttributeUsage(AttributeTargets.Method | AttributeTargets.Delegate)]
+    public class PathItemTypeAttribute : Attribute
+    {
+        public PathItemType Type { get; }
+        public PathItemTypeAttribute(PathItemType type) => Type = type;
+    }
+}
